backend/requirements.txt
Flask>=2.1
Flask-Login>=0.6
Flask-Cors>=3.0
Flask-SQLAlchemy>=3.0
python-dotenv>=0.21.0
